Wrap redigo Pool with waitPool.
The main feature of waitPool is adding blocking timeout when
getting redis connection from the pool.

Default behaviour of redigo Pool is to simply returns invalid connection
when the pool is full (no available connection).

With waitPool, we wait for the configured timeout before returning
the invalid connection.

From super simple test in localhost, using max_active=20
- without wait pool: only able to handle 70 concurrent goroutines
- with wait pool, 1s timeout: 490 concurrent goroutines.
The number will be lower in production environment